rename handle module to mempool_interface_impl
