+++ mempool/src/lib.rs
@@ -25,7 +25,7 @@ pub use interface::MempoolInterface;
 
 use crate::config::GetMemoryUsage;
 use crate::error::Error as MempoolError;
-use crate::handle::MempoolInterfaceImpl;
+use crate::mempool_interface_impl::MempoolInterfaceImpl;
 use crate::pool::Mempool;
 
 pub use crate::pool::SystemClock;
@@ -34,8 +34,8 @@ pub use crate::pool::SystemUsageEstimator;
 mod config;
 pub mod error;
 mod feerate;
-mod handle;
 mod interface;
+mod mempool_interface_impl;
 mod pool;
 pub mod rpc;
 pub mod tx_accumulator;
